engine/requirements.txt
# LLM (OpenRouter via OpenAI-compatible client)
openai>=1.0

# Backend
fastapi>=0.104
uvicorn[standard]>=0.24
websockets>=12.0
aiosqlite>=0.19

# Utils
python-dotenv>=1.0
httpx>=0.25
pydantic>=2.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21

# Hedera — install separately:
# pip install git+https://github.com/hashgraph/hedera-agent-kit-py.git
